* [Android] rm useless code `WXInstanceExceptionRecord`
